huge refactor, plus docs, and misc changes
