+++ src/cli/rcfile.rs
@@ -1,0 +1,58 @@
+use super::config::{DEFAULT_RCFILE, RCFILE};
+use super::error::{Error, IoError};
+use crate::lib::model::variables::Variable;
+use colored::Colorize;
+use std::{fs, io::ErrorKind::NotFound};
+
+use super::cli::{handle_errors, handle_input};
+
+/// Load Rustcalc's rcfile from the default location, `RCFILE`.
+/// This function has side effects:
+/// * Files may be created
+/// * May write to stdout
+///
+/// ## Input
+/// * `vars` - A mutable reference to the applications variables. Executing the rcfile may create variables.
+///
+/// ## Output
+/// Returns an empty `Result` on success, or a `CliError` from io operations
+pub fn load(vars: &mut Vec<Variable>) -> Result<(), Error> {
+    let path = match RCFILE.as_deref() {
+        Some(path) => path,
+        None => {
+            return Err(IoError::new(NotFound, "Couldn't get path for config directory").into())
+        }
+    };
+
+    // If RCFile doesn't exist, create it and write the default contents
+    if !path.exists() {
+        println!(
+            "RCFile doesn't exist. Creating default at [{}]",
+            path.to_string_lossy()
+        );
+        fs::write(path, DEFAULT_RCFILE)?;
+    }
+
+    // Read
+    let lines = fs::read_to_string(path)?;
+
+    // Filter out empty and comment lines
+    let lines = lines
+        .lines()
+        .enumerate()
+        .filter(|(_, l)| !(l.trim().is_empty() || l.starts_with("//")));
+
+    // Feed each line through `handle_input` and make use of `handle_errors`
+    // Succesfully executing statements are silent
+    for (n, line) in lines {
+        if let Err(inner) = handle_input(line, vars) {
+            let message = handle_errors(inner, line);
+            println!(
+                "Error in RCFile on line [{}]: {}",
+                format!("{}", n).red(),
+                message
+            );
+        }
+    }
+    Ok(())
+}
