+++ src/lib/eval.rs
@@ -1,12 +1,15 @@
 use super::model::{constants::Constant, errors::Error, operators::Operator, tokens::Token};
 
+/// Evaluate a list of tokens
+/// * `tokens` - The tokens
+///
+/// Returns the result as a 64-bit float or an `Error`
 pub fn eval(tokens: &[Token]) -> Result<f64, Error> {
     // We need a mutable copy of the tokens
     let mut stack: Vec<Token> = tokens.iter().rev().cloned().collect();
     let mut args: Vec<f64> = Vec::new();
 
-    while !stack.is_empty() {
-        let token = stack.pop().unwrap();
+    while let Some(token) = stack.pop() {
 
         match token {
             Token::Number { value } => {
@@ -37,7 +40,7 @@ pub fn eval(tokens: &[Token]) -> Result<f64, Error> {
     }
 
     // Result
-    if stack.is_empty() && args.len() == 1 {
+    if args.len() == 1 {
         return Ok(args[0]);
     }
     Err(Error::EmptyStack)
