+++ src/lib/model/mod.rs
@@ -1,6 +1,7 @@
 pub mod constants;
 pub mod errors;
 pub mod operators;
-mod representable;
 pub mod tokens;
 pub mod variables;
+
+mod representable;
